DeviceInfo: drop providerId, introduce providerName
There were a bad design in DeviceInfo type that was still involving a .providerId that dynamically could changes based on getEnv
this information have no sense to be in a hw getDeviceInfo because that's a dynamic value, it's not a immutable/raw info

This commit introduces back the feature via a getProviderId that takes DeviceInfo and gives you that value
We also introduces helpers to implement a "blue" dot on the manager tab
