+++ src/__tests__/hw/getLatestFirmwareForDevice.js
@@ -16,7 +16,7 @@ test("1.2.0", async () => {
     version: "1.2",
     mcuVersion: "1.0",
     majMin: "1.2",
-    providerId: 1,
+    providerName: null,
     targetId: 823132162,
     isOSU: false,
     isBootloader: false,
@@ -33,7 +33,7 @@ test("1.3.1", async () => {
     version: "1.3.1",
     mcuVersion: "1.1",
     majMin: "1.3",
-    providerId: 1,
+    providerName: null,
     targetId: 823132162,
     isOSU: false,
     isBootloader: false,
@@ -52,7 +52,7 @@ test("1.4.2", async () => {
     managerAllowed: false,
     mcuVersion: "1.6",
     pinValidated: true,
-    providerId: 1,
+    providerName: null,
     majMin: "1.4",
     targetId: 823132163,
   };
@@ -69,7 +69,7 @@ test("1.5.5", async () => {
     managerAllowed: false,
     mcuVersion: "1.7",
     pinValidated: false,
-    providerId: 1,
+    providerName: null,
     majMin: "1.5",
     targetId: 823132164
   };
@@ -87,7 +87,7 @@ test("nano x 1.1.6", async () => {
     isOSU: false,
     managerAllowed: false,
     pinValidated: true,
-    providerId: 1,
+    providerName: null,
     targetId: 855638020,
   };
   const res = await manager.getLatestFirmwareForDevice(deviceInfo);
@@ -99,7 +99,7 @@ test("nano x 1.2.4-1", async () => {
     version: "1.2.4-1",
     mcuVersion: "2.8",
     majMin: "1.2",
-    providerId: 1,
+    providerName: null,
     targetId: 855638020,
     isOSU: false,
     isBootloader: false,
@@ -118,7 +118,7 @@ test("nanoS das", async () => {
     managerAllowed: false,
     mcuVersion: "1.5",
     pinValidated: true,
-    providerId: 2,
+    providerName: "das",
     targetId: 823132163,
     version: "1.4.2-das",
   };
