+++ src/hw/getDeviceInfo.js
@@ -5,15 +5,7 @@ import { log } from "@ledgerhq/logs";
 import type Transport from "@ledgerhq/hw-transport";
 import getVersion from "./getVersion";
 import type { DeviceInfo } from "../types/manager";
-import { getEnv } from "../env";
-
-const PROVIDERS: { [_: string]: number } = {
-  "": 1,
-  das: 2,
-  club: 3,
-  shitcoins: 4,
-  ee: 5,
-};
+import { PROVIDERS } from "../manager/provider";
 
 const ManagerAllowedFlag = 0x08;
 const PinValidatedFlag = 0x80;
@@ -27,12 +19,8 @@ export default async function getDeviceInfo(
   const isOSU = seVersion.includes("-osu");
   const version = seVersion.replace("-osu", "");
   const m = seVersion.match(/([0-9]+.[0-9]+)(.[0-9]+)?(-(.*))?/);
-  const [, majMin, , , providerName] = m || [];
-  const forceProvider = getEnv("FORCE_PROVIDER");
-  const providerId =
-    forceProvider && forceProvider !== 1
-      ? forceProvider
-      : PROVIDERS[providerName] || 1;
+  const [, majMin, , , postDash] = m || [];
+  const providerName = PROVIDERS[postDash] ? postDash : null;
   const isBootloader = (targetId & 0xf0000000) !== 0x30000000;
   const flag = flags.length > 0 ? flags[0] : 0;
   const managerAllowed = !!(flag & ManagerAllowedFlag);
@@ -49,7 +37,7 @@ export default async function getDeviceInfo(
     version,
     mcuVersion,
     majMin,
-    providerId,
+    providerName: providerName || null,
     targetId,
     isOSU,
     isBootloader,
